Refactored file structure of lytenyte core
