+++ v1-migration-folder/packages/lytenyte-core/src/__tests__/grid-filtered.stories.tsx
@@ -1,21 +1,21 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { Header } from "../header/header";
 import { HeaderRow } from "../header/header-row";
-import { Root } from "../root";
+import { Root } from "../root/root";
 import { RowsContainer } from "../rows/rows";
 import { Viewport } from "../viewport/viewport";
-import { useLyteNyte } from "../../state/use-lytenyte";
+import { useLyteNyte } from "../state/use-lytenyte";
 import { useId } from "react";
 import { HeaderCell } from "../header/header-cell";
-import type { Column } from "../../+types";
+import type { Column } from "../+types";
 import { HeaderGroupCell } from "../header/header-group-cell";
-import { useClientRowDataSource } from "../../row-data-source/use-client-data-source";
+import { useClientRowDataSource } from "../row-data-source/use-client-data-source";
 import { RowsBottom, RowsCenter, RowsTop } from "../rows/rows-sections";
-import { RowHandler } from "./sample-data/row-handler";
 import { bankData } from "./sample-data/bank-data";
+import { RowHandler } from "./sample-data/row-handler";
 
 const meta: Meta = {
-  title: "Grid/Pinned Columns",
+  title: "Grid/Filtering",
 };
 
 export default meta;
@@ -40,78 +40,11 @@ const columns: Column<any>[] = [
   { id: "y" },
 ];
 
-const columnsPinStart: Column<any>[] = [
-  { id: "age" },
-  { id: "job", pin: "start" },
-  { id: "balance", pin: "start" },
-  { id: "education" },
-  { id: "marital" },
-  { id: "default" },
-  { id: "housing" },
-  { id: "loan" },
-  { id: "contact" },
-  { id: "day" },
-  { id: "month" },
-  { id: "duration" },
-  { id: "campaign" },
-  { id: "pdays" },
-  { id: "previous" },
-  { id: "poutcome" },
-  { id: "y" },
-];
-const columnsPinEnd: Column<any>[] = [
-  { id: "age" },
-  { id: "job", pin: "end" },
-  { id: "balance", pin: "end" },
-  { id: "education" },
-  { id: "marital" },
-  { id: "default" },
-  { id: "housing" },
-  { id: "loan" },
-  { id: "contact" },
-  { id: "day" },
-  { id: "month" },
-  { id: "duration" },
-  { id: "campaign" },
-  { id: "pdays" },
-  { id: "previous" },
-  { id: "poutcome" },
-  { id: "y" },
-];
-
-const columnsPinBoth: Column<any>[] = [
-  { id: "age" },
-  { id: "job", pin: "end" },
-  { id: "balance", pin: "end" },
-  { id: "education" },
-  { id: "marital" },
-  { id: "default" },
-  { id: "housing" },
-  { id: "loan" },
-  { id: "contact" },
-  { id: "day" },
-  { id: "month" },
-  { id: "duration", pin: "start" },
-  { id: "campaign", pin: "start" },
-  { id: "pdays" },
-  { id: "previous" },
-  { id: "poutcome" },
-  { id: "y" },
-];
-
-const columnsPinEndOnly: Column<any>[] = [
-  { id: "job", pin: "end" },
-  { id: "balance", pin: "end" },
-];
-const columnsPinStartOnly: Column<any>[] = [
-  { id: "job", pin: "start" },
-  { id: "balance", pin: "start" },
-];
-
 function Component({ data = bankData }: { data?: any[] }) {
   const ds = useClientRowDataSource({
     data: data,
   });
+
   const g = useLyteNyte({
     gridId: useId(),
     columns,
@@ -126,17 +59,26 @@ function Component({ data = bankData }: { data?: any[] }) {
         <button onClick={() => g.state.rtl.set((prev) => !prev)}>
           RTL: {g.state.rtl.get() ? "Yes" : "No"}
         </button>
+        <button
+          onClick={() => {
+            g.state.filterModel.set([
+              { field: "job", kind: "string", value: "unemployed", operator: "equals" },
+            ]);
+          }}
+        >
+          Filtered
+        </button>
+        <button
+          onClick={() => {
+            g.state.filterModel.set([]);
+          }}
+        >
+          No Filtered
+        </button>
       </div>
-      <div>
-        <button onClick={() => g.state.columns.set(columnsPinStart)}>Pin Start</button>
-        <button onClick={() => g.state.columns.set(columnsPinEnd)}>Pin End</button>
-        <button onClick={() => g.state.columns.set(columnsPinBoth)}>Pin Both</button>
-        <button onClick={() => g.state.columns.set(columns)}>No Pin</button>
-        <button onClick={() => g.state.columns.set(columnsPinEndOnly)}>Pin End Only</button>
-        <button onClick={() => g.state.columns.set(columnsPinStartOnly)}>Pin Start Only</button>
-      </div>
+      <div></div>
 
-      <div style={{ width: "100%", height: "95vh", border: "1px solid black" }}>
+      <div style={{ width: "100%", height: "90vh", border: "1px solid black" }}>
         <Root grid={g}>
           <Viewport>
             <Header>
@@ -156,6 +98,25 @@ function Component({ data = bankData }: { data?: any[] }) {
                       return (
                         <HeaderCell
                           cell={c}
+                          onClick={() => {
+                            const current = g.api.sortForColumn(c.column.id);
+
+                            if (current == null) {
+                              g.state.sortModel.set([
+                                {
+                                  columnId: c.column.id,
+                                  sort: { kind: "string" },
+                                  isDescending: false,
+                                },
+                              ]);
+                              return;
+                            }
+                            if (!current.sort.isDescending) {
+                              g.state.sortModel.set([{ ...current.sort, isDescending: true }]);
+                            } else {
+                              g.state.sortModel.set([]);
+                            }
+                          }}
                           key={c.column.id}
                           style={{ border: "1px solid black", background: "lightgray" }}
                         />
@@ -184,6 +145,6 @@ function Component({ data = bankData }: { data?: any[] }) {
   );
 }
 
-export const PinnedColumns: StoryObj = {
+export const Filtering: StoryObj = {
   render: Component,
 };
