+++ v1-migration-folder/packages/lytenyte-core/src/__tests__/grid-navigation.css
@@ -1,0 +1,4 @@
+*:focus {
+  background-color: gainsboro;
+  border: 2px solid blue;
+}
