+++ v1-migration-folder/packages/lytenyte-core/src/__tests__/grid-pinned-columns.stories.tsx
@@ -1,21 +1,21 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { Header } from "../header/header";
 import { HeaderRow } from "../header/header-row";
-import { Root } from "../root";
+import { Root } from "../root/root";
 import { RowsContainer } from "../rows/rows";
 import { Viewport } from "../viewport/viewport";
-import { useLyteNyte } from "../../state/use-lytenyte";
+import { useLyteNyte } from "../state/use-lytenyte";
 import { useId } from "react";
 import { HeaderCell } from "../header/header-cell";
-import type { Column } from "../../+types";
+import type { Column } from "../+types";
 import { HeaderGroupCell } from "../header/header-group-cell";
-import { useClientRowDataSource } from "../../row-data-source/use-client-data-source";
+import { useClientRowDataSource } from "../row-data-source/use-client-data-source";
 import { RowsBottom, RowsCenter, RowsTop } from "../rows/rows-sections";
 import { RowHandler } from "./sample-data/row-handler";
 import { bankData } from "./sample-data/bank-data";
 
 const meta: Meta = {
-  title: "Grid/Normal Rendering",
+  title: "Grid/Pinned Columns",
 };
 
 export default meta;
@@ -40,6 +40,74 @@ const columns: Column<any>[] = [
   { id: "y" },
 ];
 
+const columnsPinStart: Column<any>[] = [
+  { id: "age" },
+  { id: "job", pin: "start" },
+  { id: "balance", pin: "start" },
+  { id: "education" },
+  { id: "marital" },
+  { id: "default" },
+  { id: "housing" },
+  { id: "loan" },
+  { id: "contact" },
+  { id: "day" },
+  { id: "month" },
+  { id: "duration" },
+  { id: "campaign" },
+  { id: "pdays" },
+  { id: "previous" },
+  { id: "poutcome" },
+  { id: "y" },
+];
+const columnsPinEnd: Column<any>[] = [
+  { id: "age" },
+  { id: "job", pin: "end" },
+  { id: "balance", pin: "end" },
+  { id: "education" },
+  { id: "marital" },
+  { id: "default" },
+  { id: "housing" },
+  { id: "loan" },
+  { id: "contact" },
+  { id: "day" },
+  { id: "month" },
+  { id: "duration" },
+  { id: "campaign" },
+  { id: "pdays" },
+  { id: "previous" },
+  { id: "poutcome" },
+  { id: "y" },
+];
+
+const columnsPinBoth: Column<any>[] = [
+  { id: "age" },
+  { id: "job", pin: "end" },
+  { id: "balance", pin: "end" },
+  { id: "education" },
+  { id: "marital" },
+  { id: "default" },
+  { id: "housing" },
+  { id: "loan" },
+  { id: "contact" },
+  { id: "day" },
+  { id: "month" },
+  { id: "duration", pin: "start" },
+  { id: "campaign", pin: "start" },
+  { id: "pdays" },
+  { id: "previous" },
+  { id: "poutcome" },
+  { id: "y" },
+];
+
+const columnsPinEndOnly: Column<any>[] = [
+  { id: "job", pin: "end" },
+  { id: "balance", pin: "end" },
+];
+const columnsPinStartOnly: Column<any>[] = [
+  { id: "job", pin: "start" },
+  { id: "balance", pin: "start" },
+];
+
 function Component({ data = bankData }: { data?: any[] }) {
   const ds = useClientRowDataSource({
     data: data,
@@ -58,6 +126,14 @@ function Component({ data = bankData }: { data?: any[] }) {
         <button onClick={() => g.state.rtl.set((prev) => !prev)}>
           RTL: {g.state.rtl.get() ? "Yes" : "No"}
         </button>
+      </div>
+      <div>
+        <button onClick={() => g.state.columns.set(columnsPinStart)}>Pin Start</button>
+        <button onClick={() => g.state.columns.set(columnsPinEnd)}>Pin End</button>
+        <button onClick={() => g.state.columns.set(columnsPinBoth)}>Pin Both</button>
+        <button onClick={() => g.state.columns.set(columns)}>No Pin</button>
+        <button onClick={() => g.state.columns.set(columnsPinEndOnly)}>Pin End Only</button>
+        <button onClick={() => g.state.columns.set(columnsPinStartOnly)}>Pin Start Only</button>
       </div>
 
       <div style={{ width: "100%", height: "95vh", border: "1px solid black" }}>
@@ -108,13 +184,6 @@ function Component({ data = bankData }: { data?: any[] }) {
   );
 }
 
-export const JustScrollableRows: StoryObj = {
+export const PinnedColumns: StoryObj = {
   render: Component,
-};
-export const EmptyDataset: StoryObj = {
-  render: () => <Component data={[]} />,
-};
-
-export const MassiveDataset: StoryObj = {
-  render: () => <Component data={[...bankData, ...bankData, ...bankData, ...bankData]} />,
 };
