+++ v1-migration-folder/packages/lytenyte-core/src/__tests__/grid-spanning.stories.tsx
@@ -1,0 +1,129 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { Header } from "../header/header";
+import { HeaderRow } from "../header/header-row";
+import { Root } from "../root/root";
+import { RowsContainer } from "../rows/rows";
+import { Viewport } from "../viewport/viewport";
+import { useLyteNyte } from "../state/use-lytenyte";
+import { useId, useState } from "react";
+import { HeaderCell } from "../header/header-cell";
+import type { Column } from "../+types";
+import { HeaderGroupCell } from "../header/header-group-cell";
+import { useClientRowDataSource } from "../row-data-source/use-client-data-source";
+import { RowsBottom, RowsCenter, RowsTop } from "../rows/rows-sections";
+import { RowHandler } from "./sample-data/row-handler";
+import { bankData } from "./sample-data/bank-data";
+
+const meta: Meta = {
+  title: "Grid/Spanning",
+};
+
+export default meta;
+
+const columns: Column<any>[] = [
+  { id: "age", rowSpan: 2, colSpan: 3 },
+  { id: "job" },
+  { id: "balance" },
+  { id: "education" },
+  { id: "marital" },
+  { id: "default" },
+  { id: "housing", colSpan: (r) => (r.colIndex % 2 ? 3 : 1), pin: "end" },
+  { id: "loan", pin: "end", width: 100 },
+  { id: "contact" },
+  { id: "day" },
+  { id: "month" },
+  { id: "duration" },
+  { id: "campaign" },
+  { id: "pdays" },
+  { id: "previous" },
+  { id: "poutcome" },
+  { id: "y" },
+];
+
+function Component({ data = bankData }: { data?: any[] }) {
+  const [d, setData] = useState(() => data);
+  const [pt, setPt] = useState<any[]>(() => []);
+  const [bt, setBt] = useState<any[]>(() => []);
+
+  const ds = useClientRowDataSource({
+    data: d,
+    topData: pt,
+    bottomData: bt,
+  });
+
+  const g = useLyteNyte({
+    gridId: useId(),
+    columns,
+    rowDataSource: ds,
+  });
+
+  const view = g.view.useValue();
+
+  return (
+    <div>
+      <div>
+        <button onClick={() => g.state.rtl.set((prev) => !prev)}>
+          RTL: {g.state.rtl.get() ? "Yes" : "No"}
+        </button>
+      </div>
+      <div>
+        <button onClick={() => setData(data)}>Large</button>
+        <button onClick={() => setData([])}>Empty</button>
+        <button onClick={() => setData(data.slice(3, 5))}>Tiny</button>
+        <button onClick={() => setPt(data.slice(0, 2))}>Pin Two Top</button>
+        <button onClick={() => setPt([])}>Remove Two Top</button>
+        <button onClick={() => setBt(data.slice(8, 10))}>Pin Two Bot</button>
+        <button onClick={() => setBt([])}>Remove Two Bot</button>
+      </div>
+
+      <div style={{ width: "100%", height: "90vh", border: "1px solid black" }}>
+        <Root grid={g}>
+          <Viewport>
+            <Header>
+              {view.header.layout.map((row, i) => {
+                return (
+                  <HeaderRow headerRowIndex={i} key={i}>
+                    {row.map((c) => {
+                      if (c.kind === "group") {
+                        return (
+                          <HeaderGroupCell
+                            cell={c}
+                            key={c.idOccurrence}
+                            style={{ border: "1px solid black", background: "lightgray" }}
+                          />
+                        );
+                      }
+                      return (
+                        <HeaderCell
+                          cell={c}
+                          key={c.column.id}
+                          style={{ border: "1px solid black", background: "lightgray" }}
+                        />
+                      );
+                    })}
+                  </HeaderRow>
+                );
+              })}
+            </Header>
+
+            <RowsContainer>
+              <RowsTop>
+                <RowHandler rows={view.rows.top} />
+              </RowsTop>
+              <RowsCenter>
+                <RowHandler rows={view.rows.center} />
+              </RowsCenter>
+              <RowsBottom>
+                <RowHandler rows={view.rows.bottom} />
+              </RowsBottom>
+            </RowsContainer>
+          </Viewport>
+        </Root>
+      </div>
+    </div>
+  );
+}
+
+export const Spanning: StoryObj = {
+  render: Component,
+};
