+++ v1-migration-folder/packages/lytenyte-core/src/cells/cell-default.tsx
@@ -1,4 +1,4 @@
-import type { CellRendererParams } from "../../+types";
+import type { CellRendererParams } from "../+types";
 
 export function CellDefault<T>({ grid, column, row }: CellRendererParams<T>) {
   const r = grid.api.fieldForColumn(column, row);
