+++ v1-migration-folder/packages/lytenyte-core/src/header/header-cell.tsx
@@ -1,45 +1,50 @@
 import { forwardRef, useMemo, type CSSProperties, type JSX } from "react";
-import type { HeaderGroupCellLayout } from "../../+types";
+import type { HeaderCellLayout } from "../+types";
 import { useGridRoot } from "../context";
-import { getTranslate } from "@1771technologies/lytenyte-shared";
+import { getTranslate, sizeFromCoord } from "@1771technologies/lytenyte-shared";
 import { fastDeepMemo } from "@1771technologies/lytenyte-react-hooks";
 
-export interface HeaderGroupCellProps {
-  readonly cell: HeaderGroupCellLayout;
+export interface HeaderCellProps<T> {
+  readonly cell: HeaderCellLayout<T>;
 }
 
-const HeaderGroupCellImpl = forwardRef<
+const HeaderCellImpl = forwardRef<
   HTMLDivElement,
-  JSX.IntrinsicElements["div"] & HeaderGroupCellProps
+  JSX.IntrinsicElements["div"] & HeaderCellProps<any>
 >(function HeaderCell({ cell, children, ...props }, forwarded) {
   const ctx = useGridRoot().grid.state;
 
   const xPositions = ctx.xPositions.useValue();
-  const height = ctx.headerGroupHeight.useValue();
 
-  const x = xPositions[cell.colStart];
-  const width = xPositions[cell.colEnd] - x;
+  const width = sizeFromCoord(cell.colStart, xPositions, cell.colSpan);
+  const rowSpan = cell.rowEnd - cell.rowStart;
 
   const isSticky = !!cell.colPin;
   const viewport = ctx.viewportWidthInner.useValue();
+  const rtl = ctx.rtl.useValue();
 
   const styles = useMemo(() => {
     const styles: CSSProperties = {};
     if (isSticky) {
       styles.position = "sticky";
-      styles.left = 0;
+
+      if (rtl) styles.right = 0;
+      else styles.left = 0;
+
       styles.zIndex = 11;
     }
 
     if (cell.colPin === "end") {
       const spaceLeft = xPositions.at(-1)! - xPositions[cell.colStart];
-      styles.transform = getTranslate(viewport - spaceLeft, 0);
-    } else {
-      styles.transform = getTranslate(xPositions[cell.colStart], 0);
-    }
 
+      const x = viewport - spaceLeft;
+      styles.transform = getTranslate(rtl ? -x : x, 0);
+    } else {
+      const x = xPositions[cell.colStart];
+      styles.transform = getTranslate(rtl ? -x : x, 0);
+    }
     return styles;
-  }, [cell.colPin, cell.colStart, isSticky, viewport, xPositions]);
+  }, [cell.colPin, cell.colStart, isSticky, rtl, viewport, xPositions]);
 
   return (
     <div
@@ -49,16 +54,17 @@ const HeaderGroupCellImpl = forwardRef<
       style={{
         ...props.style,
         ...styles,
-        gridRow: "1 / 2",
+        gridRowStart: 1,
+        gridRowEnd: rowSpan + 1,
         gridColumn: "1 / 2",
         width,
-        height,
+        height: "100%",
         boxSizing: "border-box",
       }}
     >
-      {children == undefined ? cell.id : children}
+      {children == undefined ? cell.column.id : children}
     </div>
   );
 });
 
-export const HeaderGroupCell = fastDeepMemo(HeaderGroupCellImpl);
+export const HeaderCell = fastDeepMemo(HeaderCellImpl);
