+++ v1-migration-folder/packages/lytenyte-core/src/header/header-group-cell.tsx
@@ -1,50 +1,45 @@
 import { forwardRef, useMemo, type CSSProperties, type JSX } from "react";
-import type { HeaderCellLayout } from "../../+types";
+import type { HeaderGroupCellLayout } from "../+types";
 import { useGridRoot } from "../context";
-import { getTranslate, sizeFromCoord } from "@1771technologies/lytenyte-shared";
+import { getTranslate } from "@1771technologies/lytenyte-shared";
 import { fastDeepMemo } from "@1771technologies/lytenyte-react-hooks";
 
-export interface HeaderCellProps<T> {
-  readonly cell: HeaderCellLayout<T>;
+export interface HeaderGroupCellProps {
+  readonly cell: HeaderGroupCellLayout;
 }
 
-const HeaderCellImpl = forwardRef<
+const HeaderGroupCellImpl = forwardRef<
   HTMLDivElement,
-  JSX.IntrinsicElements["div"] & HeaderCellProps<any>
+  JSX.IntrinsicElements["div"] & HeaderGroupCellProps
 >(function HeaderCell({ cell, children, ...props }, forwarded) {
   const ctx = useGridRoot().grid.state;
 
   const xPositions = ctx.xPositions.useValue();
+  const height = ctx.headerGroupHeight.useValue();
 
-  const width = sizeFromCoord(cell.colStart, xPositions, cell.colSpan);
-  const rowSpan = cell.rowEnd - cell.rowStart;
+  const x = xPositions[cell.colStart];
+  const width = xPositions[cell.colEnd] - x;
 
   const isSticky = !!cell.colPin;
   const viewport = ctx.viewportWidthInner.useValue();
-  const rtl = ctx.rtl.useValue();
 
   const styles = useMemo(() => {
     const styles: CSSProperties = {};
     if (isSticky) {
       styles.position = "sticky";
-
-      if (rtl) styles.right = 0;
-      else styles.left = 0;
-
+      styles.left = 0;
       styles.zIndex = 11;
     }
 
     if (cell.colPin === "end") {
       const spaceLeft = xPositions.at(-1)! - xPositions[cell.colStart];
-
-      const x = viewport - spaceLeft;
-      styles.transform = getTranslate(rtl ? -x : x, 0);
+      styles.transform = getTranslate(viewport - spaceLeft, 0);
     } else {
-      const x = xPositions[cell.colStart];
-      styles.transform = getTranslate(rtl ? -x : x, 0);
+      styles.transform = getTranslate(xPositions[cell.colStart], 0);
     }
+
     return styles;
-  }, [cell.colPin, cell.colStart, isSticky, rtl, viewport, xPositions]);
+  }, [cell.colPin, cell.colStart, isSticky, viewport, xPositions]);
 
   return (
     <div
@@ -54,17 +49,16 @@ const HeaderCellImpl = forwardRef<
       style={{
         ...props.style,
         ...styles,
-        gridRowStart: 1,
-        gridRowEnd: rowSpan + 1,
+        gridRow: "1 / 2",
         gridColumn: "1 / 2",
         width,
-        height: "100%",
+        height,
         boxSizing: "border-box",
       }}
     >
-      {children == undefined ? cell.column.id : children}
+      {children == undefined ? cell.id : children}
     </div>
   );
 });
 
-export const HeaderCell = fastDeepMemo(HeaderCellImpl);
+export const HeaderGroupCell = fastDeepMemo(HeaderGroupCellImpl);
