+++ v1-migration-folder/packages/lytenyte-core/src/state/helpers/empty-row-data-source.ts
@@ -4,4 +4,5 @@ export const emptyRowDataSource: RowDataSource<any> = {
   init: () => {},
   rowById: () => null,
   rowByIndex: () => null,
+  rowExpand: () => {},
 };
